add zstack installation test for installing rabbitmq server in remote host.

+++ integrationtest/vm/installation/test_stub.py
@@ -38,11 +38,10 @@ def check_str(string):
         return ""
     return string
 
-def execute_shell_in_process(cmd, tmp_file):
+def execute_shell_in_process(cmd, tmp_file, timeout = 3600, no_timeout_excep = False):
     logfd = open(tmp_file, 'w', 0)
     process = subprocess.Popen(cmd, executable='/bin/sh', shell=True, stdout=logfd, stderr=logfd, universal_newlines=True)
 
-    timeout = 3600
     start_time = time.time()
     while process.poll() is None:
         curr_time = time.time()
@@ -53,8 +52,11 @@ def execute_shell_in_process(cmd, tmp_file):
             logfd = open(tmp_file, 'r')
             test_util.test_logger('[shell:] %s [timeout logs:] %s' % (cmd, '\n'.join(logfd.readlines())))
             logfd.close()
-            os.system('rm -f %s' % tmp_file)
-            test_util.test_fail('[shell:] %s timeout, after %d seconds' % (cmd, timeout))
+            if no_timeout_excep:
+                test_util.test_logger('[shell:] %s timeout, after %d seconds' % (cmd, timeout))
+            else:
+                os.system('rm -f %s' % tmp_file)
+                test_util.test_fail('[shell:] %s timeout, after %d seconds' % (cmd, timeout))
         print('Installation: %d' % int(test_time))
         time.sleep(1)
     logfd.close()
@@ -102,7 +104,18 @@ def execute_all_install(ssh_cmd, target_file, tmp_file):
     if process_result != 0:
         cmd = '%s "cat /tmp/zstack_installation.log"' % ssh_cmd
         execute_shell_in_process(cmd, tmp_file)
-        test_util.test_fail('zstack installation failed')
+        if 'no management-node-ready message received within' in open(tmp_file).read():
+            times = 12
+            cmd = '%s "zstack-ctl status"' % ssh_cmd
+            while (times > 0):
+                time.sleep(10)
+                process_result = execute_shell_in_process(cmd, tmp_file, 10, True)
+                if process_result == 0:
+                    test_util.test_logger("management node start after extra %d seconds" % (10 - times + 1) * 10 )
+                    return 0
+                test_util.test_logger("mn node is still not started up, wait for another 10 seconds...")
+            else:
+                test_util.test_fail('zstack installation failed')
 
 def only_install_zstack(ssh_cmd, target_file, tmp_file):
     env_var = "WEBSITE='%s'" % 'localhost'
